requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cmake==3.17.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
lxml==4.5.1
numpy==1.19.0
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
queuelib==1.5.0
requests==2.24.0
Scrapy==2.2.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
Twisted==20.3.0
union==0.1.10
urllib3==1.25.9
w3lib==1.22.0
zope.interface==5.1.0
